[SDK-109] Http communication has been outsourced to its own module in the project
