+++ .gitignore
@@ -11,3 +11,4 @@ njams4-sdk-root.iml
 njams-sdk/njams4-sdk.iml
 njams-sdk-communication-cloud/njams4-sdk-communication-cloud.iml
 njams-sdk-sample-client/njams4-sdk-sample-client.iml
+njams-sdk-communication-http/njams4-sdk-communication-http.iml
