+++ njams-sdk-communication-http/.gitignore
@@ -1,0 +1,3 @@
+/target/
+/.classpath
+/.project
