+++ njams-sdk-communication-http/pom.xml
@@ -1,0 +1,39 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+
+    <groupId>com.faizsiegeln.njams</groupId>
+    <artifactId>njams4-sdk-communication-http</artifactId>
+    <name>njams4-sdk-communication-http</name>
+
+    <parent>
+        <groupId>com.faizsiegeln</groupId>
+        <artifactId>njams4-sdk-root</artifactId>
+        <version>4.0.6-SNAPSHOT</version>
+    </parent>
+
+<dependencies>
+
+    <!-- provided - not shaded -->
+    <dependency>
+        <groupId>com.faizsiegeln.njams</groupId>
+        <artifactId>njams4-sdk</artifactId>
+        <version>4.0.6-SNAPSHOT</version>
+        <scope>provided</scope>
+    </dependency>
+
+    <dependency>
+        <groupId>org.slf4j</groupId>
+        <artifactId>slf4j-api</artifactId>
+        <scope>provided</scope>
+    </dependency>
+
+    <dependency>
+        <groupId>junit</groupId>
+        <artifactId>junit</artifactId>
+        <scope>test</scope>
+    </dependency>
+</dependencies>
+</project>
