+++ njams-sdk-communication-http/src/main/java/com.im.njams.sdk.communication.http/HttpConstants.java
