+++ njams-sdk-communication-http/src/main/resources/com.im.njams.sdk.communication.connectable.Receiver
@@ -1,0 +1,2 @@
+com.im.njams.sdk.communication.http.connectable.HttpReceiver
+com.im.njams.sdk.communication.http.https.connectable.HttpsSender
